pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["yazan <yazan.alshekha@outlook.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
djangorestframework = "^3.12.0"
psycopg2-binary = "^2.8.6"
djangorestframework_simplejwt = "^4.4.0"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
